requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.15
aiosignal==1.3.2
asyncio==3.4.3
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
frozenlist==1.5.0
idna==3.10
iniconfig==2.1.0
multidict==6.3.1
packaging==24.2
pluggy==1.5.0
propcache==0.3.1
pycparser==2.22
pytest==8.3.5
pytest-aiohttp==1.1.0
pytest-asyncio==0.26.0
python-qpid-proton==0.39.0
requests==2.32.3
urllib3==2.3.0
yarl==1.18.3
